爬取所有资源

+++ src/main/java/cn/sw/project/service/ArticleService.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Map;
 
 public interface ArticleService {
-    int addArticle(Article article);
+    long addArticle(Article article);
 
     List<Article> getArticleList(Map<String, Object> params);
 }
