+++ src/main/java/cn/sw/project/util/LuceneUtil.java
@@ -97,7 +97,7 @@ public class LuceneUtil {
 	
 	public static Article doucmentToArticle(Document document){
 		Article article=new Article();
-		article.setId(Integer.valueOf(document.get("Id")));
+		article.setId(Long.valueOf(document.get("Id")));
 		article.setAuthor(document.get("author"));
 		article.setUrl(document.get("link"));
 		article.setContent(document.get("content"));
